requirements.txt
fastapi[standard]
uvicorn[standard]
python-multipart
mysql-connector-python
python-jose[cryptography]
bcrypt==3.2.0
passlib==1.7.4
